autorizacija nastavak

+++ piklon-users/src/main/java/ba/team12/piklonusers/controllers/UserController.java
@@ -58,12 +58,15 @@ public class UserController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public String deleteNote(@PathVariable(value = "id") Long id) throws ChangeSetPersister.NotFoundException {
+    public String deleteNote(@RequestHeader Token  tokenadminov,@PathVariable(value = "id") Long id) throws ChangeSetPersister.NotFoundException {
         User user = userRepository.findById(id).orElseThrow(() -> new FileSystemNotFoundException("User not found"));
-        Token token = tokenRepository.findByUserId(id);
-        userRepository.delete(user);
-        tokenRepository.delete(token);
-        return "User is deleted";
+       if(tokenadminov.getUser_token().contains("admin")) {
+           Token tokenuserov = tokenRepository.findByUserId(id);
+           userRepository.delete(user);
+           tokenRepository.delete(tokenuserov);
+           return "User is deleted";
+       }
+       return "failed";
     }
 
     /*@PostMapping("/login")
